refactor: move getArticleEntries logic out of vivliostyle.config.js
Agent-Logs-Url: https://github.com/ymm-oss/daigirin-template/sessions/5f0d8864-508b-4d9f-bcb5-7dc96d3c29c4
